display-drivers: Disable xiaomi layer handling
Since the xiaomi fod/aod layer flags are never set on AOSP, this code
is useless for us. Furthermore the missing AOD layer flag was always
leading to the driver thinking that we are exiting the AOD and thus it
switched to NOLP mode (which, on some devices, leads to a black screen
if the brightness isn't updated afterwards). Without the layers, we need
to restore the last brightness after setting nolp mode (what was
originally done in mi_sde_connector_update_aod_status).

Change-Id: I6ea2ba6960308056a945c781572519e4e59d45b5
Signed-off-by: Fiqri Ardyansyah <fiqri191002@gmail.com>